% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idrm.R
\name{corIDR}
\alias{corIDR}
\title{Irreproducible Discovery Rate analysis: Replicate Correlation Metric}
\usage{
corIDR(x, mu = 1, sigma = 0.5, rho = 0.5, p = 0.5, ...)
}
\arguments{
\item{x}{an n by m numeric matrix, where m = num of replicates, n = num of
observations. Numerical values representing the significance of the
observations. Note that significant signals are expected to have large
values of x. In case that smaller values represent higher significance
(e.g. p-value), a monotonic transformation needs to be applied to reverse
the order before using this function, for example, -log(p-value).}

\item{mu}{a starting value for the scalar mean for the reproducible
component.}

\item{sigma}{a starting value for the scalar standard deviation (diagonal
covariance) of the reproducible component.}

\item{rho}{a starting value for the scalar correlation coefficient
(off-diagonal correlation) of the reproducible component.}

\item{p}{a starting value for the proportion of the reproducible component.}

\item{...}{additional arguments passed to \code{\link[scider]{est.IDRm}}.}
}
\value{
a matrix of estimated correlation coefficients (NA if fit fails).
}
\description{
Compute reproducible correlation coefficient over replicates.
}
\examples{
data("simu.idr",package = "idr")
# simu.idr$x and simu.idr$y are p-values
# Transfer them such that large values represent significant ones
x <- cbind(-simu.idr$x, -simu.idr$y)

mu <- 2.6
sigma <- 1.3
rho <- 0.8
p <- 0.7

print(corIDR(x))

}
