% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idrm.R
\name{e.step.2mnormal}
\alias{e.step.2mnormal}
\title{E-step for parameterized multivariate 2-component Gaussian mixture
models}
\usage{
e.step.2mnormal(z, mu, sigma, rho, p)
}
\arguments{
\item{z}{data matrix}

\item{mu}{scalar mean for the reproducible component.}

\item{sigma}{scalar standard deviation (diagonal covariance) of the
reproducible component.}

\item{rho}{scalar correlation coefficient (off-diagonal correlation) of the
reproducible component.}

\item{p}{mixing proportion of the reproducible component.}
}
\value{
e.z a numeric vector, where each entry represents the estimated
  expected conditional probability that an observation is in the
  reproducible component.
}
\description{
Expectation step in the EM algorithm for parameterized multivariate
2-component Gaussian mixture models with (1-p)N(0, 1, 0) + pN(mu, sigma,
rho).
Code adapted from \code{\link[idr]{e.step.2normal}}.
}
\examples{
z.1 <- c(rnorm(500, 0, 1), rnorm(500, 3, 1))
rho <- 0.8

## The component with higher values is correlated with corr. coefficient=0.8
z.2 <- c(rnorm(500, 0, 1), rnorm(500, 3 + 0.8*(z.1[501:1000]-3), (1-rho^2)))

## Starting values
mu0 <- 3
sigma0 <- 1
rho0 <- 0.85
p0 <- 0.55

e.z <- e.step.2mnormal(cbind(z.1,z.2), mu0, sigma0, rho0, p0)
}
\keyword{internal}
