% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idrm.R
\name{est.IDRm.sample}
\alias{est.IDRm.sample}
\title{Irreproducible Discovery Rate analysis with Sub-Sampling}
\usage{
est.IDRm.sample(x, mu, sigma, rho, p, frac = 0.7, nsamp = 100,
  verbose = FALSE, plot = FALSE, ...)
}
\arguments{
\item{x}{an n by m numeric matrix, where m = num of replicates, n = num of
observations. Numerical values representing the significance of the
observations. Note that significant signals are expected to have large
values of x. In case that smaller values represent higher significance
(e.g. p-value), a monotonic transformation needs to be applied to reverse
the order before using this function, for example, -log(p-value).}

\item{mu}{a starting value for the scalar mean for the reproducible
component.}

\item{sigma}{a starting value for the scalar standard deviation (diagonal
covariance) of the reproducible component.}

\item{rho}{a starting value for the scalar correlation coefficient
(off-diagonal correlation) of the reproducible component.}

\item{p}{a starting value for the proportion of the reproducible component.}

\item{frac}{fraction of observations chosen in each sample. Default 0.7.}

\item{nsamp}{number of samples. Default 100.}

\item{verbose}{If TRUE, print helpful messages. Default FALSE.}

\item{plot}{If TRUE, plot summary figures. Default FAlSE.}

\item{...}{additional arguments passed to \code{\link[scider]{est.IDRm}}.}
}
\value{
a list with the following elements: \itemize{ \item{mean_para}{ mean
  estimated parameters: p, rho, mu, sigma.}\item{para_bp}{ 
  \code{\link[graphics]{boxplot}} summary for estimated parameters over 
  samples.} \item{mean_idr}{ a numeric vector of mean local idr for each 
  observation (i.e. estimated conditional probablility for each observation 
  to belong to the irreproducible component.} \item{sd_idr}{ a numeric 
  vector of s.d. of local idr for each observation.} \item{IDR}{ a numerical
  vector of the expected irreproducible discovery rate for observations that
  are as irreproducible or more irreproducible than the given observations.}
  \item{times_sampled}{ a numerical vector of counts for each time an
  observation was included in a sample.} \item{num_failed}{ number of times
  a sampled fit failed for any reason.} }
}
\description{
Fit a multivariate Gaussian copula mixture model to multiple sub-samples of
observations.
}
\examples{
data("simu.idr",package = "idr")
# simu.idr$x and simu.idr$y are p-values
# Transfer them such that large values represent significant ones
x <- cbind(-simu.idr$x, -simu.idr$y)

mu <- 2.6
sigma <- 1.3
rho <- 0.8
p <- 0.7

idr.out <- est.IDRm.sample(x, mu, sigma, rho, p, nsamp = 5)

plot(-log10(idr.out$IDR),idr.out$sd_idr)
abline(v = 2, col = "red", lty = 2)

}
