% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gayoData}
\alias{gayoData}
\alias{gayoSCDC}
\alias{gayoBulkTLR}
\title{Data: Normalized RNA-Seq data and meta data for Gayo et al. Human DC Study}
\description{
Data: Normalized RNA-Seq data and meta data for Gayo et al. Human DC Study
}
\details{
Normalized RSEM transcriptome abundance estimates for 393 cells
  passing sample filter. Normalized RSEM log(TPM+1) estimates as described
  in SOM of Gayo et al. Metadata and alignment quality control metrics for
  RSEM transcriptome alignments are included in the SummarizedExperiment
  object.

Average normalized RSEM transcriptome abundance estimates for 5
  human bulk TLR perturbations. Normalized RSEM log(TPM+1) estimates as
  described in SOM of Gayo et al.
}
\examples{
data(gayoSCDC)
data(gayoBulkTLR)
}
