% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get.repro.tresh.adjacency}
\alias{get.repro.tresh.adjacency}
\title{Reproducible Module analysis: Molecule Adjacency Matrix}
\usage{
get.repro.tresh.adjacency(x, r, mad_constant = 1.4826, var_thresh = 10^-5,
  pair_pthresh = 0.1)
}
\arguments{
\item{x}{an n by m numeric matrix, where m = num of samples, n = num of 
molecules. Numerical values representing the expression level of the 
molecule.}

\item{r}{factor. Sample replicate identity.}

\item{mad_constant}{scale factor passed to \code{\link[stats]{mad}} in 
scaling z-transformed correlation coefficients.}

\item{var_thresh}{numeric. Molecules are only analyzed if their variance is 
greater than this threshold in all replicate groups.}

\item{pair_pthresh}{numeric. Molecule pairs are called "reproducibly 
correlated" if, across all replicate groups, their z-value corresponds to 
a two-sided p-value below this threshold.}
}
\value{
Sparse adjacency matrix (\code{\link[Matrix]{dgCMatrix-class}}) with
  logical is_variable attribute flagging genes included in analysis.
}
\description{
Generate molecule-molecule adjacency matrix for molecule pairs exhibiting 
reproducible correlations across patients.
}
\examples{
library("SummarizedExperiment")
data("fluidigm",package = "scRNAseq")

x = assay(fluidigm)[sample(x = seq_len(nrow(fluidigm)),size = 1000),]
r = factor(fluidigm$Coverage_Type)
A = get.repro.tresh.adjacency(x,r)

library("igraph")
g = graph.adjacency(A, mode = "undirected")

hist(degree(g))
}
