% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{getMarkers}
\alias{getMarkers}
\title{Reproducible Marker Prediction}
\usage{
getMarkers(x, r, g, r_de = r, g_target = NULL, external_markers = NULL,
  de_idr_thresh = 0.01, module_bonf_thresh = 0.01, var_thresh = 10^-5,
  mad_constant = 1.4826, pair_pthresh = 0.01, idr_mu = 2, idr_sigma = 2,
  idr_rho = 0.5, idr_p = 0.01, plot_venn = FALSE, ...)
}
\arguments{
\item{x}{an n by m numeric matrix, where m = num of samples, n = num of 
molecules. Numerical values representing the expression level of the 
molecule.}

\item{r}{factor. Sample replicate identity.}

\item{g}{factor. Sample cluster identity. NA samples are ignored in de.}

\item{r_de}{factor. If specified, use this argument as de analysis sample
replicate identity instead of \code{r}.}

\item{g_target}{character. If not null, only include genes up-regulated in
this cluster.}

\item{external_markers}{character. If not NULL, only return markers if they
are included in this argument.}

\item{de_idr_thresh}{numeric. IDR threshold for calling differential
expression (default 0.01).}

\item{module_bonf_thresh}{numeric. Bonferroni-adjusted threshold for calling
reproducible module genes (default 0.01).}

\item{var_thresh}{numeric. Molecules are only analyzed if their variance is 
greater than this threshold in all replicate groups.}

\item{mad_constant}{scale factor passed to \code{\link[stats]{mad}} in 
scaling z-transformed correlation coefficients.}

\item{pair_pthresh}{numeric. Molecule pairs are called "reproducibly 
correlated" if, across all replicate groups, their z-value corresponds to 
a two-sided p-value below this threshold.}

\item{idr_mu}{a starting value for idr mean (see 
\code{\link[scider]{est.IDRm}}).}

\item{idr_sigma}{a starting value for the idr standard deviation (see 
\code{\link[scider]{est.IDRm}}).}

\item{idr_rho}{a starting value for the idr correlation coefficient (see 
\code{\link[scider]{est.IDRm}}).}

\item{idr_p}{a starting value for the proportion of the reproducible tests 
(see \code{\link[scider]{est.IDRm}}).}

\item{plot_venn}{if TRUE, plot venn diagram of marker candidates.}

\item{...}{additional arguments passed to \code{\link[scider]{est.IDRm}}.}
}
\value{
list. Genes in de, modules, and (optional) external marker sets, as
  well as intersection.
}
\description{
Wrapper function for differential expression (de) analysis and reproducible
module (module) analysis.
}
\examples{
library("SummarizedExperiment")
data("fluidigm",package = "scRNAseq")

x = assay(fluidigm)[sample(x = seq_len(nrow(fluidigm)),size = 1000),]
r = factor(fluidigm$Coverage_Type)
g = factor(fluidigm$Biological_Condition)
getMarkers(x,r,g)

}
