% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idrm.R
\name{kruskalIDRm}
\alias{kruskalIDRm}
\title{Irreproducible Discovery Rate analysis: Differential Expression}
\usage{
kruskalIDRm(x, r, g, var_thresh = 10^-5, idr_mu = 2, idr_sigma = 2,
  idr_rho = 0.5, idr_p = 0.01)
}
\arguments{
\item{x}{an n by m numeric matrix, where m = num of samples, n = num of
molecules. Numerical values representing the expression level of the
molecule.}

\item{r}{factor. Sample replicate identity.}

\item{g}{factor. Sample cluster identity, passed to
\code{\link[stats]{kruskal.test}}.}

\item{var_thresh}{numeric. Genes are only analyzed if their variance is
greater than this threshold in all replicate groups.}

\item{idr_mu}{a starting value for idr mean (see
\code{\link[scrap]{est.IDRm}}).}

\item{idr_sigma}{a starting value for the idr standard deviation (see
\code{\link[scrap]{est.IDRm}}).}

\item{idr_rho}{a starting value for the idr correlation coefficient (see
\code{\link[scrap]{est.IDRm}}).}

\item{idr_p}{a starting value for the proportion of the reproducible tests
(see \code{\link[scrap]{est.IDRm}}).}
}
\value{
a list with the following elements: \itemize{ \item{idr}{ list.
  \code{\link[scrap]{est.IDRm}} output.}\item{kruskal_pvals}{ matrix of
  Kruskal test p-values used in IDR analysis.} \item{is_replicated}{
  logical. TRUE if gene passes variance filter.} }
}
\description{
Perform differential expression analysis (using
\code{\link[stats]{kruskal.test}}) over multiple replicate experiments,
followed by irreproducible discovery rate (IDR) analysis.
}
\examples{
library("SummarizedExperiment")
data("fluidigm",package = "scRNAseq")

x = assay(fluidigm)[sample(x = seq_len(nrow(fluidigm)),size = 1000),]
r = factor(fluidigm$Coverage_Type)
g = factor(fluidigm$Biological_Condition)
kruskal_idrm_obj = kruskalIDRm(x,r,g)

# Compare significance of 2 coverage types and color by reproducibility
plot(-log10(kruskal_idrm_obj$kruskal_pvals)[kruskal_idrm_obj$is_replicated,],
     main = "-log10(p-value) by Coverage Type",
     xlim = c(0,max(-log10(kruskal_idrm_obj$kruskal_pvals),na.rm = TRUE)),
     ylim = c(0,max(-log10(kruskal_idrm_obj$kruskal_pvals),na.rm = TRUE)),
     col = 1 + (kruskal_idrm_obj$idr$IDR < 0.01), pch = 16)
abline(0,1, lty = 2)

}
