% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{kruskalMeta}
\alias{kruskalMeta}
\title{Simple Meta-Analysis: Differential Expression}
\usage{
kruskalMeta(x, r, g, var_thresh = 10^-5)
}
\arguments{
\item{x}{an n by m numeric matrix, where m = num of samples, n = num of 
molecules. Numerical values representing the expression level of the 
molecule.}

\item{r}{factor. Sample replicate identity.}

\item{g}{factor. Sample cluster identity, passed to 
\code{\link[stats]{kruskal.test}}.}

\item{var_thresh}{numeric. Molecules are only analyzed if their variance is 
greater than this threshold in all replicate groups.}
}
\value{
a list with the following elements: \itemize{ \item{meta_pvals}{ 
  Stouffer p-values.} \item{FDR}{ FDR q-values.} \item{kruskal_pvals}{ 
  matrix of Kruskal test p-values used in meta-analysis.} 
  \item{is_replicated}{ logical. TRUE if molecule passes variance filter.} }
}
\description{
Perform differential expression analysis (using 
\code{\link[stats]{kruskal.test}}) over multiple replicate experiments, 
followed by Stouffer p-value combination and FDR analysis.
}
\examples{
library("SummarizedExperiment")
data("fluidigm",package = "scRNAseq")

x = assay(fluidigm)[sample(x = seq_len(nrow(fluidigm)),size = 1000),]
r = factor(fluidigm$Coverage_Type)
g = factor(fluidigm$Biological_Condition)
kruskal_meta_obj = kruskalMeta(x,r,g)

# Compare significance of 2 coverage types and color by reproducibility
plot(
  -log10(kruskal_meta_obj$kruskal_pvals)[kruskal_meta_obj$is_replicated, ],
  main = "-log10(p-value) by Coverage Type",
  xlim = c(0, max(
    -log10(kruskal_meta_obj$kruskal_pvals), na.rm = TRUE
  )),
  ylim = c(0, max(
    -log10(kruskal_meta_obj$kruskal_pvals), na.rm = TRUE
  )),
  col = 1 +
    (kruskal_meta_obj$FDR[kruskal_meta_obj$is_replicated] < 0.01),
  pch = 16
)
abline(0, 1, lty = 2)

}
