% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{ksMeta}
\alias{ksMeta}
\title{Simple Meta-Analysis: Signature Comparison}
\usage{
ksMeta(x, x0, r, var_thresh = 10^-5, ...)
}
\arguments{
\item{x}{an n by m numeric matrix, where m = num of samples, n = num of 
signatures. Numerical values representing signature values.}

\item{x0}{reference signature distribution.}

\item{r}{factor. Sample replicate identity.}

\item{var_thresh}{numeric. Molecules are only analyzed if their variance is 
greater than this threshold in all replicate groups.}

\item{...}{Additional arguments passed to \code{\link[stats]{ks.test}}).}
}
\value{
a list with the following elements: \itemize{ \item{meta_pvals}{ 
  Stouffer p-values.} \item{FDR}{ FDR q-values.} \item{ks_pvals}{ 
  matrix of KS-test p-values used in meta-analysis.} 
  \item{is_replicated}{ logical. TRUE if molecule passes variance filter.} }
}
\description{
Compare distributions of signature values to a reference distribution (using
\code{\link[stats]{ks.test}}) over multiple replicate experiments, 
followed by Stouffer p-value combination and FDR analysis.
}
\examples{
x = matrix(rnorm(1000),nrow = 10)
x0 = rnorm(1000)
r = factor(rep(1:2,each = 50))
ks_meta_obj = ksMeta(x = x,x0 = x0,r)

# Compare significance of 2 coverage types and color by reproducibility
plot(-log10(ks_meta_obj$ks_pvals)[ks_meta_obj$is_replicated,],
     main = "-log10(p-value) by Replicate",
     xlim = c(0,max(-log10(ks_meta_obj$ks_pvals),na.rm = TRUE)),
     ylim = c(0,max(-log10(ks_meta_obj$ks_pvals),na.rm = TRUE)),
     col = 1 + (ks_meta_obj$FDR[ks_meta_obj$is_replicated] < 0.01),
     pch = 16)
abline(0,1, lty = 2)

}
