% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idrm.R
\name{loglik.2mnormal}
\alias{loglik.2mnormal}
\title{Compute log-likelihood of parameterized multivariate 2-component Gaussian
mixture models}
\usage{
loglik.2mnormal(z, mu, sigma, rho, p)
}
\arguments{
\item{z}{data matrix}

\item{mu}{scalar mean for the reproducible component.}

\item{sigma}{scalar standard deviation (diagonal covariance) of the
reproducible component.}

\item{rho}{scalar correlation coefficient (off-diagonal correlation) of the
reproducible component.}

\item{p}{mixing proportion of the reproducible component.}
}
\value{
Log-likelihood of the multivariate 2-component Gaussian mixture
  models (1-p)N(0, 1, 0) + pN(mu, sigma, rho).
}
\description{
Compute the log-likelihood for parameterized multivariate 2-component
Gaussian mixture models with (1-p)N(0, 1, 0) + pN(mu, sigma, rho).
}
\examples{
z.1 <- c(rnorm(500, 0, 1), rnorm(500, 3, 1))
rho <- 0.8

## The component with higher values is correlated with corr. coefficient=0.8
z.2 <- c(rnorm(500, 0, 1), rnorm(500, 3 + 0.8*(z.1[501:1000]-3), (1-rho^2)))

## Starting values
mu <- 3
sigma <- 1
rho <- 0.85
p <- 0.55

## The function is currently defined as
loglik <- loglik.2mnormal(cbind(z.1,z.2), mu, sigma, rho, p)

loglik
}
\keyword{internal}
