% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{pzipdegree}
\alias{pzipdegree}
\title{Reproducible Module analysis: Zero-Inflated Poisson Degree Distribution}
\usage{
pzipdegree(g)
}
\arguments{
\item{g}{an undirected igraph graph (see
\code{\link[igraph]{graph.adjacency}}).}
}
\value{
For each node, the upper tail probability under the fitted Poisson
  component distribution.
}
\description{
Models degree distribution of an undirected graph by a zero-inflated Poisson
distribution, using the fitted Poisson parameters to define the null for
upper-tail probabilities.
}
\examples{
library("SummarizedExperiment")
data("fluidigm",package = "scRNAseq")

x = assay(fluidigm)[sample(x = seq_len(nrow(fluidigm)),size = 1000),]
r = factor(fluidigm$Coverage_Type)
A = get.repro.thresh.adjacency(x,r)

library("igraph")
g = graph.adjacency(A, mode = "undirected")

pdegree = pzipdegree(g)
plot(degree(g),-log10(p.adjust(pdegree,method = "bonferroni")))
}
